Transfered moving into the spawner, Implemented difficulty
Changed so that the moving objects are moved by the spawner and get reused instead of spawning new objects.
This allowed me to easier set up difficulty, so i started with a linear kind, but then though about animation curves and managed to implement curves to set up difficulty over the playtime.