Near misses destroying random barrel and give points
Implemented raycast that runs when the player moves seeing if it's closer than a variable point to then report a near miss. Near misses are reported to the spawner, leading to the destruction of a random barrel. gonna want to maybe make the barrel that's missed be imune to deletion.