Getting the full loop working
Got the start menu, loading into the game, showing a menu at the end, and then being able to either restart or go back to the menu. meant to get the highscore showing but am having a few issues. might have time to look into it tomorrow unless the school event takes up all the time.